Refactor roll call/cluster formation
